changed from league to league season, minor improvements
Changed all references of Liga methods to LigaSaison, that can be
unique for every season.
The overall change makes possible for separate seasons of the same
league to have different configurations, like if a league increases the
number of teams or introduces Relegation, or to decide for each season
if kader files are provided.
Renamed method to make camel-case.